Implement ReignEngine.BuildModel
Instance motions are updated with this function.
